+++ AppLovin/CHANGELOG.md
@@ -5,6 +5,7 @@
          * Add support for using a cached `configuration` NSDictionary for initialization.
          * Do not consider banner ad display failure as ad load failure.
          * Do not consider users closing rewarded ad early or reward validation failure as an ad load failure.
+         * Bumped AppLovin SDK plugin version to "MoPub-6.1.4.2".
    
    * 6.1.4.1
      * **Note**: This version is only compatible with the 5.5.0+ release of the MoPub SDK.